% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_bio_prop.R
\name{calc_bio_prop}
\alias{calc_bio_prop}
\title{Biomass Proportion
Calculates the mean biomass proportion from an atlantis run}
\usage{
calc_bio_prop(atl.dir, fgs.file, timeRange)
}
\arguments{
\item{atl.dir}{character, directory containing atlantis output files}

\item{fgs.file}{character, path to the functional groups file}

\item{timeRange}{numeric vector, range of years to summarize}
}
\value{
vector of mean biomass proportions for each group in the specified time range
}
\description{
Biomass Proportion
Calculates the mean biomass proportion from an atlantis run
}
