% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_cum_bio.R
\name{calc_cum_bio}
\alias{calc_cum_bio}
\title{Cumulative Biomass Distribution
Calculates the cumulative biomass statistics from Libralato et al, 2019}
\usage{
calc_cum_bio(bio.df, tl.df, show.plot = F, out.dir)
}
\arguments{
\item{tl.df}{dataframe. Trophic level dataframe with the trophic levels of each group (species x year x TL) returned from est_trophic_level_time}

\item{show.plot}{logical. whether it should return plots as well as data}

\item{out.dir}{string. Output directory to save the results}

\item{biomass.df}{dataframe. Biomass dataframe with the biomass of each group (species x year x biomass) returned from est_biomass_time}
}
\value{
dataframe of steepness, biomass at inflection, and trophic level at inflection for each year, as well as fitted parameters
}
\description{
Cumulative Biomass Distribution
Calculates the cumulative biomass statistics from Libralato et al, 2019
}
