% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_divergence.R
\name{calc_divergence}
\alias{calc_divergence}
\title{Divergence metrics for species composition
Calculates the compositional divergence metrics from atlantis output}
\usage{
calc_divergence(bio.df, atl.dir, ref.prop, fgs.file, show.plot = F)
}
\arguments{
\item{bio.df}{dataframe. Atlantis biomass output dataframe with columns: Time, Code, Biomass}

\item{ref.prop}{dataframe. Reference proportion dataframe with the reference proportions of each group (species x prop) returned from calc_ref_prop}

\item{show.plot}{logical. whether it should return plots as well as data}
}
\value{
dataframe of divergence metrics over time
}
\description{
Divergence metrics for species composition
Calculates the compositional divergence metrics from atlantis output
}
