% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_fish_prod.R
\name{calc_fish_prod}
\alias{calc_fish_prod}
\title{Fish productivity indicators
Function to calculate smallfish-per-largefish ratio based on Atlantis output}
\usage{
calc_fish_prod(
  param.dir,
  atl.dir,
  out.dir,
  survdat,
  show.plot = FALSE,
  timeRange
)
}
\arguments{
\item{param.dir}{Character String. Path to Parameter directory}

\item{atl.dir}{Character String. Path to output directory}

\item{out.dir}{character string. Path to output file}

\item{survdat}{Dataframe. Survey data with columns: Code, AGE, LENGTH}

\item{show.plot}{logical. If TRUE, will plot the results}

\item{timeRange}{numeric vector. Time range to filter the data (e.g., c(1980, 2020)). If NULL, uses all available data.}
}
\description{
Fish productivity indicators
Function to calculate smallfish-per-largefish ratio based on Atlantis output
}
