% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_foodweb.R
\name{calc_foodweb}
\alias{calc_foodweb}
\title{Food web indices
calculates food web indices from detailed diet output}
\usage{
calc_foodweb(
  atl.dir,
  fgs.file,
  param.dir,
  dietSource,
  timeRange,
  show.plot = F,
  figure.dir = NA,
  out.dir = NA,
  bySpecies = FALSE,
  ...
)
}
\arguments{
\item{atl.dir}{character string. Path to Atlantis output directory}

\item{fgs.file}{Character String. Name of groups.csv file}

\item{param.dir}{character string. Path to Atlantis parameter directory}

\item{dietSource}{Character String. Whether to use realized diets (diet), detailedDiet (detdiet) or parameter files (param)}

\item{figure.dir}{character string. Path to directory for figures}

\item{out.dir}{character string. Path to output directory for plots}

\item{bySpecies}{logical. If TRUE, outputs a dataframe with metrics calculated for each species/node. Defaults to FALSE.}

\item{plot}{logical. whether it should return plots as well as data}
}
\value{
list of metric timeseries, metric status, food web, and ecosystem status. If bySpecies is TRUE, returns a single dataframe of metrics.
}
\description{
Food web indices
calculates food web indices from detailed diet output
}
