% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodweb_functions.R
\name{calculate_avg_jaccard_weighted}
\alias{calculate_avg_jaccard_weighted}
\title{Weighted Jaccard Similarity (Ružička Similarity)
Function to calculate the average weighted Jaccard similarity}
\usage{
calculate_avg_jaccard_weighted(g)
}
\arguments{
\item{g}{igraph object. directed food web graph}
}
\value{
numeric average weighted jaccard similarity value
}
\description{
Weighted Jaccard Similarity (Ružička Similarity)
Function to calculate the average weighted Jaccard similarity
}
