% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodweb_functions.R
\name{calculate_coherence}
\alias{calculate_coherence}
\title{Coherence
Function to calculate trophic coherence}
\usage{
calculate_coherence(adj_matrix, trophic_levels)
}
\arguments{
\item{adj_matrix}{matrix. directed flow adjacency matrix of the food web with names}

\item{trophic_levels}{numeric. named vector of trophic levels by species in the adj_matrix returned from calculate_trophic_levels_corrected()}
}
\value{
numeric coherence value
}
\description{
Coherence
Function to calculate trophic coherence
}
