% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_soe_thresholds.R
\name{create_soe_thresholds}
\alias{create_soe_thresholds}
\title{Replicate Link and Watson study using NEUS region}
\usage{
create_soe_thresholds(epu)
}
\arguments{
\item{epu}{A character string indicating the EPU to use. eg. MAB, GB, GOM}
}
\description{
Use regional Primary production to create new thresholds. This comes from Kim Hyde
}
