% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_bmsy.R
\name{est_bmsy}
\alias{est_bmsy}
\title{Estimates Bmsy from fitting a shaeffer model using biomass and catch time series}
\usage{
est_bmsy(biomass, catch, start_r = 0.5, start_k = NULL)
}
\arguments{
\item{biomass}{numeric vector of biomass values}

\item{catch}{numeric vector of catch values}

\item{start_r}{numeric starting value for r in the non-linear least squares fit (default is 0.5)}

\item{start_k}{numeric starting value for carrying capacity K. If NULL, it will be set to max(biomass) * 1.5.}
}
\value{
A list containing: distance and closest point
}
\description{
Estimates Bmsy from fitting a shaeffer model using biomass and catch time series
}
\details{
If a dimension is unbounded, you can use -Inf for lower bounds and Inf for upper bounds.
}
