% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_trophic_level_from_params.r
\name{est_trophic_level_from_params}
\alias{est_trophic_level_from_params}
\title{Estimate Trophic level from diet matrix}
\usage{
est_trophic_level_from_params(
  prm_biol =
    "C:/Users/andrew.beet/Documents/myWork/githubRepos/neus-atlantis/currentVersion/at_biology.prm",
  fgs =
    "C:/Users/andrew.beet/Documents/myWork/githubRepos/neus-atlantis/currentVersion/neus_groups.csv",
  plottl = F
)
}
\arguments{
\item{prm_biol}{Character String. Path to biology prm file (Atlantis)}

\item{fgs}{Character String. Name of groups.csv file}

\item{plottl}{Boolean. Plot the Trophic level to window. Default = F}
}
\value{
list
\item{trophiclevel}{dataframe. Species, year, trophic level}
\item{figuretl}{ggplot2 object. trophic level bar plot by species}
}
\description{
uses PPrey values to determine Trophic level
}
\details{
Need to process detailed diet data frirst
linux command line
}
\section{zip up file}{
gzip -k neus_outputDetailedDietCheck.txt
}

\section{then remove zeros and save ans another zip}{
zcat neus_outputDetailedDietCheck.txt.gz | awk 'NR > 1{s=0; for (i=6;i<=NF;i++) s+=$i; if (s!=0)print}' | gzip > DetDiet.gz
}

\section{strip headers from original file}{
zcat neus_outputDetailedDietCheck.txt.gz | head -n1 | gzip > DetDietHead.gz
}

\section{concatenate them}{
cat DetDietHead.gz DetDiet.gz > DetDiet2.gz
}

