% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_trophic_level_time.r
\name{est_trophic_level_time}
\alias{est_trophic_level_time}
\title{Estimate Trophic level over time from detailed diet}
\usage{
est_trophic_level_time(
  param.dir =
    "C:/Users/andrew.beet/Documents/myWork/githubRepos/neus-atlantis/currentVersion",
  atl.dir = here::here("other/dev_3a75e57d1"),
  fgs = "neus_groups.csv",
  detDietfile = "DetDiet2.gz",
  timeRange = 1988:2017,
  plottl = F
)
}
\arguments{
\item{atl.dir}{Character String. Path to output directory}

\item{fgs}{Character String. Name of groups.csv file}

\item{detDietfile}{Character String. Name of processed zipped detailed diet file}

\item{timeRange}{Vector}

\item{plottl}{Boolean. Plot the Trophic level to window. Default = F}

\item{param, dir}{Character String. Path to Parameter directory}
}
\value{
list
\item{trophiclevel}{dataframe. Species, year, trophic level}
\item{figuretl}{ggplot2 object. trophic level plots over time by species}
}
\description{
Need to process detailed diet data first
linux command line
}
\section{zip up file}{
gzip -k neus_outputDetailedDietCheck.txt
}

\section{then remove zeros and save ans another zip}{
zcat neus_outputDetailedDietCheck.txt.gz | awk 'NR > 1{s=0; for (i=6;i<=NF;i++) s+=$i; if (s!=0)print}' | gzip > DetDiet.gz
}

\section{strip headers from original file}{
zcat neus_outputDetailedDietCheck.txt.gz | head -n1 | gzip > DetDietHead.gz
}

\section{concatenate them}{
cat DetDietHead.gz DetDiet.gz > DetDiet2.gz
}

