% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prm_functions.R
\name{get_age_mat}
\alias{get_age_mat}
\title{Atlantis Age at Maturity
Function to retrieve age at maturity parameters}
\usage{
get_age_mat(bio.file, write = F)
}
\arguments{
\item{bio.file}{character. Path to biomass.prm file}

\item{write}{logical. whether to write output to file}
}
\value{
dataframe of species and age at maturity
}
\description{
Atlantis Age at Maturity
Function to retrieve age at maturity parameters
}
