% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bmsy_atl.R
\name{get_bmsy_atl}
\alias{get_bmsy_atl}
\title{Generates a range of ecological indicators from atlantis output to be used as a summary of possible states}
\usage{
get_bmsy_atl(param.dir, atl.dir)
}
\arguments{
\item{param.dir}{Character String. Path to Parameter directory}

\item{atl.dir}{Character String. Path to output directory}
}
\value{
dataframe of ecological indicators over time
}
\description{
Generates a range of ecological indicators from atlantis output to be used as a summary of possible states
}
