% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bmsy_ss.R
\name{get_bmsy_ss}
\alias{get_bmsy_ss}
\title{Function to get stockSMART Bmsy for species and supply a default if not found}
\usage{
get_bmsy_ss(fgs, default.bmsy.frac)
}
\arguments{
\item{fgs}{Character String. Path to the functional groups file}

\item{default.bmsy.frac}{Numeric. Default fraction of biomass to use if Bmsy is not found in stockSMART}
}
\value{
data.frame. Species, Bmsy
}
\description{
Function to get stockSMART Bmsy for species and supply a default if not found
}
