% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prm_functions.R
\name{get_param_FSPB}
\alias{get_param_FSPB}
\title{Edit Atlantis FSPB
Function to retrieve FSPB paramters}
\usage{
get_param_FSPB(bio.prm)
}
\arguments{
\item{bio.file}{character. Path to biomass.prm file}

\item{write}{logical. whether to write output to file}
}
\value{
dataframe of species and FSPB
}
\description{
Edit Atlantis FSPB
Function to retrieve FSPB paramters
}
