% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pp.R
\name{get_pp}
\alias{get_pp}
\title{Read in the Phyto_Forcing_xxx.nc files to get estimates of PP}
\usage{
get_pp(bgm, pathToForcing)
}
\arguments{
\item{bgm}{Path to bgm file}

\item{pathToForcing}{Character sting. Path to location of forcing files}

\item{convertNtoC}{Boolean. Convert Nitrogen (Atlantis units) to Carbon (Default = T)}
}
\value{
list of output
\item{annual}{annual primary production (tons Nitrogen or Carbon) for all primary producers in NEUS}
\item{daily}{daily primary production (tons Nitrogen or Carbon)}
\item{dailyspecies}{daily primary production (tons Nitrogen or Carbon)}
\item{dailybox}{daily primary production by box (tons Nitrogen or Carbon)}
\item{totalArea}{area domain}
}
\description{
Get primary production data from nc forcing files
}
