% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ppc.R
\name{get_ppc}
\alias{get_ppc}
\title{Read in the output to get primary production consumed}
\usage{
get_ppc(param.dir, atl.dir, fgs, dietSource, timeRange, plottl = F)
}
\arguments{
\item{param.dir}{Character String. Path to Parameter directory}

\item{atl.dir}{Character String. Path to output directory}

\item{fgs}{Character String. Name of groups.csv file}

\item{dietSource}{string. Source of diet matrix, either 'diet','detDiet', or 'prm'. Default is 'prm'.}

\item{timeRange}{Vector}

\item{plottl}{Boolean. Plot the Trophic level to window. Default = F}
}
\value{
data.frame
}
\description{
Get primary production data from nc forcing files
NEUS atlantis is forced with 3 primary producer species (Diatoms, Dinoflagellates,Picophytoplankton)
}
