% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_eco_indicators.R
\name{make_eco_indicators}
\alias{make_eco_indicators}
\title{Creates an ecological indicator dataset from atlantis output}
\usage{
make_eco_indicators(
  param.dir,
  atl.dir,
  group.index,
  timeRange,
  fgs.file,
  dietSource,
  cloud = F
)
}
\arguments{
\item{param.dir}{Character String. Path to Parameter directory}

\item{atl.dir}{Character String. Path to output directory}

\item{group.index}{Character String. Name of the group index with non-atlantis categories}

\item{timeRange}{Numeric Vector. Range of years to include in the analysis}

\item{fgs.file}{Character String. Name of the functional groups file}

\item{dietSource}{Character String. Whether to use realized diets (diet), detailedDiet (detdiet) or parameter files (param)}

\item{cloud}{Logical. If TRUE, run on cloud Default is FALSE.}
}
\value{
list
\item{trophiclevel}{dataframe. Species, year, trophic level}
\item{figuretl}{ggplot2 object. trophic level bar plot by species}
}
\description{
Creates an ecological indicator dataset from atlantis output
}
