% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_state_distance_ellipse.R
\name{make_state_distance_ellipse}
\alias{make_state_distance_ellipse}
\title{Calculate the distance from model output to desired state}
\usage{
make_state_distance_ellipse(ideal.state, ideal.center, observed.state)
}
\arguments{
\item{ideal.state}{vector. Desired state of the model defined as as range from a point (length N)}

\item{ideal.center}{vector. Center of the ideal state (length N)}

\item{observed.state}{vector. Observed state of the model defined as a range from a point (length N)}
}
\value{
data.frame
}
\description{
Calculate the distance from model output to desired state
}
