% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_state_distance_rect.R
\name{make_state_distance_rect}
\alias{make_state_distance_rect}
\title{Computes the distance between a current state and a desired state space}
\usage{
make_state_distance_rect(observed.state, desired.lower, desired.upper)
}
\arguments{
\item{observed.state}{numeric vector based on the current state of the system (where N = the number of indicators)}

\item{desired.lower}{numeric vector of lower bounds for each indicator (length N)}

\item{desired.upper}{numeric vector of upper bounds for each indicator (length N)}
}
\value{
A list containing: distance and closest point
}
\description{
Computes the distance between a current state and a desired state space
}
\details{
If a dimension is unbounded, you can use -Inf for lower bounds and Inf for upper bounds.
}
