% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_targeting_proportions.R
\name{normalize}
\alias{normalize}
\title{Normalize a Numeric Vector to Sum to 1}
\usage{
normalize(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric vector of the same length as x where the elements sum to 1.
}
\description{
This is an internal helper function that takes a numeric vector and scales its
elements so that they sum to 1.
}
