% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_eco_indicators.R
\docType{data}
\name{param.dir}
\alias{param.dir}
\title{Creates an ecological indicator dataset from atlantis output}
\format{
An object of class \code{character} of length 1.
}
\usage{
param.dir
}
\arguments{
\item{param.dir}{Character String. Path to Parameter directory}

\item{atl.dir}{Character String. Path to output directory}

\item{group.index}{Character String. Name of the group index with non-atlantis categories}

\item{fgs.file}{Character String. Name of the functional groups file}

\item{dietSource}{Character String. Whether to use realized diets (diet), detailedDiet (detdiet) or parameter files (param)}

\item{plottl}{Boolean. Plot the Trophic level to window. Default = F}
}
\value{
list
\item{trophiclevel}{dataframe. Species, year, trophic level}
\item{figuretl}{ggplot2 object. trophic level bar plot by species}
}
\description{
Creates an ecological indicator dataset from atlantis output
}
\keyword{datasets}
