% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_footprint_epu.R
\name{plot_footprint_epu}
\alias{plot_footprint_epu}
\title{plots NEUS Atlantis box structure overlaying EPU's}
\usage{
plot_footprint_epu(crs = 4269L)
}
\arguments{
\item{crs}{Integer scalar. coordinate reference system EPSG number}
}
\value{
An sf object.

A figure is also plotted to the Plots window
}
\description{
plots NEUS Atlantis box structure overlaying EPU's
}
