% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ind_pca.R
\name{plot_ind_pca}
\alias{plot_ind_pca}
\title{Plots distance metrics relative to reference run in principal component space}
\usage{
plot_ind_pca(data.dir, ref.ind.file, ref.state.file, figure.dir)
}
\arguments{
\item{data.dir}{directory where the data files are stored}

\item{ref.ind.file}{file containing the eco indicators data}

\item{ref.state.file}{file containing the reference eco state data}

\item{figure.dir}{directory where the figures will be saved}
}
\value{
dataframe of eco indicators by year
}
\description{
Plots distance metrics relative to reference run in principal component space
}
