% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodweb_functions.R
\name{calculate_avg_jaccard}
\alias{calculate_avg_jaccard}
\title{Jaccard Similarity
Function to calculate the average unweighted Jaccard similarity}
\usage{
calculate_avg_jaccard(g)
}
\arguments{
\item{g}{igraph object. directed food web graph}
}
\value{
numeric average jaccard similarity value
}
\description{
Jaccard Similarity
Function to calculate the average unweighted Jaccard similarity
}
