% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodweb_functions.R
\name{calculate_trophic_levels_corrected}
\alias{calculate_trophic_levels_corrected}
\title{Corrected Trophic Levels with assumed primary producers
Function to calculate trophic level with fixed primary producers (useful for age-structured food web)}
\usage{
calculate_trophic_levels_corrected(adj_matrix, basal_species)
}
\arguments{
\item{adj_matrix}{matrix. directed flow adjacency matrix of the food web with names}

\item{basal_species}{vector. character vector of species that are basal (primary producers)}
}
\value{
vector of trophic levels by species in the adj_matrix
}
\description{
Corrected Trophic Levels with assumed primary producers
Function to calculate trophic level with fixed primary producers (useful for age-structured food web)
}
