% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_diet_prop.R
\name{get_diet_prop}
\alias{get_diet_prop}
\title{Retreives diet proportions from parameter files}
\usage{
get_diet_prop(param.dir, atl.dir, fgs, dietFile, plottl = F, dietSource)
}
\arguments{
\item{param.dir}{Character String. Path to Parameter directory}

\item{atl.dir}{Character String. Path to output directory}

\item{fgs}{Character String. Name of groups.csv file}

\item{dietFile}{Character String. Name of processed zipped detailed diet file}

\item{plottl}{Boolean. Plot the Trophic level to window. Default = F}

\item{dietSource}{string. Source of diet matrix, either 'diet', 'detdiet' or 'prm'. Default is 'prm'.}
}
\value{
dataframe
\item{realized diet}{data.frame time,pred,prey,consumption, prop.consumption}
\item{prm diet}{pred,prey,consumption, prop.consumption}
}
\description{
Gets diet information from pPREY values in biology.prm
}
