% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pp.R
\name{get_pp}
\alias{get_pp}
\title{Read in the Phyto_Forcing_xxx.nc files to get estimates of PP}
\usage{
get_pp(bgm, pathToForcing)
}
\arguments{
\item{bgm}{Path to bgm file}

\item{pathToForcing}{Character sting. Path to location of forcing files}
}
\value{
list of output
\item{dailyspeciesbox}{daily primary production by species and atlantis box (tons Nitrogen)}
\item{dailyspecies}{daily primary production by species (tons Nitrogen)}
\item{daily}{daily primary production (tons Nitrogen)}
\item{annual}{annual primary production (tons Nitrogen) for all primary producers in NEUS}
\item{totalArea}{area domain of NEUS model (km^2)}
}
\description{
Get primary production data from nc forcing files
NEUS atlantis is forced with 3 primary producer species (Diatoms, Dinoflagellates,Picophytoplankton)
}
