% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_desired_state_distance.R
\name{make_desired_state_distance}
\alias{make_desired_state_distance}
\title{Distance fromd Desired State
Calculates the multivariate distance of observations from desired state refd on multiple atlantis runs/observations}
\usage{
make_desired_state_distance(
  param.dir,
  atl.dir,
  dietSource,
  ref.state.file,
  data.dir,
  out.dir,
  run.prefix,
  setup.file
)
}
\arguments{
\item{param.dir}{Character String. Path to Parameter directory}

\item{atl.dir}{Character String. Path to output directory}

\item{dietSource}{Character String. Whether to use realized diets (diet), detailedDiet (detdiet) or parameter files (param)}

\item{ref.state.file}{Character String. Path to ref eco state file}

\item{data.dir}{Character String. Path to directory with atlantis runs you're looking to compare}

\item{out.dir}{Character String. Path to directory where you want to save data output}

\item{run.prefix}{Character String. Prefix for the atlantis runs you're looking to compare}

\item{setup.file}{Character String. Path to setup file for the atlantis runs you're looking to compare}
}
\value{
dataframe of eco indicators by year
}
\description{
Distance fromd Desired State
Calculates the multivariate distance of observations from desired state refd on multiple atlantis runs/observations
}
