% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_eco_indicators_time.R
\name{make_eco_indicators_time}
\alias{make_eco_indicators_time}
\title{Generates a range of ecological indicators from atlantis output to be used as a summary of possible states}
\usage{
make_eco_indicators_time(
  param.dir,
  atl.dir,
  group.index,
  fgs.file,
  dietSource,
  timeRange,
  start.year = 1964,
  cloud = F,
  ref.run.dir = NULL,
  survdat.data = NA,
  debug = FALSE
)
}
\arguments{
\item{param.dir}{Character String. Path to Parameter directory}

\item{atl.dir}{Character String. Path to output directory}

\item{group.index}{Character String. Name of the group index with non-atlantis categories}

\item{fgs.file}{Character String. Name of the functional groups file}

\item{dietSource}{Character String. Whether to use realized diets (diet), detailedDiet (detdiet) or parameter files (param)}

\item{timeRange}{Numeric vector. Range of years to summarize}

\item{start.year}{Numeric. Year to start the time series. Default = 1964}

\item{cloud}{Logical. If TRUE, run on cloud Default is FALSE.}

\item{ref.run.dir}{Character String. Path to reference run directory. If NULL, no divergence metrics are calculated.}

\item{survdat.data}{dataframe of survdat length,age,weight data}

\item{debug}{Logical. If TRUE, prints progress messages. Default is FALSE.}
}
\value{
dataframe of ecological indicators over time
}
\description{
Generates a range of ecological indicators from atlantis output to be used as a summary of possible states
}
