% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_reference_state.R
\name{make_reference_state}
\alias{make_reference_state}
\title{Make Reference State
Calculates ecosystem indicators from a reference run and converts them to a desired state based on quantiles of timeseries}
\usage{
make_reference_state(
  param.dir,
  atl.dir,
  group.index,
  fgs.file,
  dietSource,
  timeRange,
  cloud = FALSE,
  desired.p = c(0.25, 0.75)
)
}
\arguments{
\item{param.dir}{Character String. Path to Parameter directory}

\item{atl.dir}{Character String. Path to output directory}

\item{group.index}{Character String. Name of the group index with non-atlantis categories}

\item{fgs.file}{Character String. Name of the functional groups file}

\item{dietSource}{Character String. Whether to use realized diets (diet), detailedDiet (detdiet) or parameter files (param)}

\item{timeRange}{Numeric Vector. Range of years to include in the analysis}

\item{cloud}{Logical. If TRUE, run on cloud Default is FALSE.}

\item{desired.p}{Numeric Vector. Quantiles to use for desired min/max values, default is c(0.25, 0.75)}

\item{out.name}{Character String. Name of the output file to save the results}
}
\value{
dataframe of eco indicators by year
}
\description{
Make Reference State
Calculates ecosystem indicators from a reference run and converts them to a desired state based on quantiles of timeseries
}
