% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distance_metrics.R, R/plot_eco_ind.R
\name{plot_distance_metrics}
\alias{plot_distance_metrics}
\title{Plot Distance Metrics
Plots distance metrics relative to reference run}
\usage{
plot_distance_metrics(figure.dir, eco.ind.file, setup.file, thresholds.file)

plot_distance_metrics(figure.dir, eco.ind.file, setup.file, thresholds.file)
}
\arguments{
\item{figure.dir}{Character String. Path to figure directory}

\item{eco.ind.file}{Character String. Path to run indicators file}

\item{setup.file}{Character String. Path to setup file}

\item{thresholds.file}{Character String. Path to thresholds file}

\item{distance.file}{Character String. Path to distance file}

\item{ref.state.file}{Character String. Path to reference file}
}
\value{
dataframe of eco indicators by year

dataframe of eco indicators by year
}
\description{
Plot Distance Metrics
Plots distance metrics relative to reference run
}
