% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_det_diet.R
\name{process_det_diet}
\alias{process_det_diet}
\title{Processes detailed diet output for use in atlantisom}
\usage{
process_det_diet(atl.dir, detDietfile, outputname, cloud = FALSE)
}
\arguments{
\item{atl.dir}{Character String. Path to output directory}

\item{detDietfile}{Character String. Name of processed zipped detailed diet file}

\item{outputname}{character String. Name of output file}

\item{cloud}{Logical. If TRUE, run on cloud Default is FALSE.}
}
\value{
DetDiet_Processed.gz
}
\description{
Processes detailed diet output for use in atlantisom
}
